Move check_user_has_role back into auth to avoid circular deps
https://github.com/radiasoft/sirepo/pull/3352#discussion_r579677950
